feat(home-screen): Resolving ToDos
