+++ lib/l10n/app_en.arb
@@ -77,5 +77,7 @@
   "favorites": "Favorites",
   "@favorites": {},
   "profile": "Profile",
-  "@profile": {}
+  "@profile": {},
+  "instructions": "Instructions",
+  "@instructions": {}
  }
