+++ lib/l10n/app_sr.arb
@@ -77,5 +77,7 @@
   "favorites": "Omiljeni treninzi",
   "@favorites": {},
   "profile": "Moj profil",
-  "@profile": {}
+  "@profile": {},
+  "instructions": "Instrukcije",
+  "@instructions": {}
 }
