+++ lib/presentations/cubit/settings/language/language_cubit.dart
@@ -21,9 +21,6 @@ class LanguageCubit extends Cubit<LanguageState> {
     }
   }
 
-  // done: This logic is ok, but orElse should not be needed
-  // Since you create the dropdown which provides supported languages inside the app
-  // there is no need to double check if the language which was sent to [changeLanguage] is supported
   Future<void> changeLanguage(String languageCode) async {
     final language = Language.supportedLanguages.firstWhere(
       (language) => language.code == languageCode,
