+++ lib/presentations/cubit/settings/language/language_state.dart
@@ -7,17 +7,3 @@ class LanguageState extends Equatable {
   @override
   List<Object> get props => [locale];
 }
-
-// done: Just have one state. No need for [LanguageChanged] extends [LanguageState]
-// just use:
-
-// class LanguageState extends Equatable {
-//   final Locale locale;
-//
-//   const LanguageChanged(this.locale);
-//
-//   @override
-//   List<Object> get props => [locale];
-// }
-
-// and then use LanguageState everywhere throughout the app.
