+++ lib/presentations/screens/home_screen_tabs/home_tab_content.dart
@@ -1,5 +1,5 @@
-import 'package:fit_lovers/data/models/exercise.dart';
 import 'package:fit_lovers/presentations/cubit/exercises/cubit/exercise_cubit.dart';
+import 'package:fit_lovers/presentations/widgets/exercise_item.dart';
 import 'package:fit_lovers/presentations/widgets/loading_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -25,52 +25,11 @@ class HomeTabContent extends StatelessWidget {
           );
         }
 
-        return Center(child: Text('No exercises available.'));
+        if (state is ExerciseError) {
+          return Center(child: Text(state.message));
+        }
+        return Container();
       },
-    );
-  }
-}
-
-class ExerciseItem extends StatelessWidget {
-  const ExerciseItem({required this.exercise, super.key});
-
-  final Exercise exercise;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(1.0),
-      child: Card(
-        color: Colors.white,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12.0),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(exercise.name!),
-              SizedBox(
-                height: 6,
-              ),
-              SizedBox(
-                height: 10,
-              ),
-              //TODO: dodaju se svi elementi iz modela
-              //TODO: Sredi cijeli UI / prevod i izgled
-
-              Text(
-                "Instructions: ${exercise.instructions}",
-                style: TextStyle(
-                  overflow: TextOverflow.ellipsis,
-                ),
-                maxLines: 3,
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
