+++ lib/presentations/screens/register_screen.dart
@@ -13,9 +13,6 @@ class RegisterScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // done: When clicking on "Register" button without accepting terms and conditions, nothing is shown to the user.
-    // done: When Register is successful no message is shown to the user.
-    // done: There is no loading state
     return Scaffold(
       appBar: CustomAppBar(
         title: (AppLocalizations.of(context)!.register),
