--- lib/presentations/widgets/consent_button.dart
@@ -1,23 +1,0 @@
-// import 'package:fit_lovers/presentations/cubit/authentication/register/register_cubit.dart';
-// import 'package:flutter/material.dart';
-// import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-
-// class ConsentButton extends StatelessWidget {
-//   const ConsentButton({
-//     super.key,
-//   });
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return CheckboxListTile(
-//       title: Text('I agree to the terms and conditions'),
-//       value: state.consent.value,
-//       onChanged: (bool? value) {
-//         if (value != null) {
-//           context.read<RegisterCubit>().consentClicked(value);
-//         }
-//       },
-//     );
-//   }
-// }
